Server
This will include server side coding which is required to connect with the client program. This program internally uses TCP protocol which is implemented in ServerSocket class.
With the help of DataInputStream it fetches the message passed by the client to it(Server). The msg will be printed after converting it to the string. This is the simplest connection program.