Client program
This code will establish connection with Socket class which is again using the same protocol i.e. TCP. We will need 2 things to connect with the server  1) IP address of server and 2) Port number of server. We are also trying to pass a msg with the help of data output stream. This msg will be reflected on the server side.
For steps refers to the .docx file in the repository.

Thank you!